Seam patches for better testing efficiency
