+++ src/agendable/web/routes/auth/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from agendable.sso.oidc.client import OidcClient
 from agendable.sso.oidc.provider import keycloak_oidc_enabled, oidc_enabled
 from agendable.web.routes.auth.router import (
     get_user_or_404,
@@ -12,15 +11,6 @@ from agendable.web.routes.auth.router import (
     render_profile_template,
     router,
 )
-
-
-def _oidc_oauth_client() -> OidcClient:
-    return oidc_oauth_client()
-
-
-def _keycloak_oidc_oauth_client() -> OidcClient:
-    return keycloak_oidc_oauth_client()
-
 
 __all__ = [
     "get_user_or_404",
