+++ src/agendable/web/routes/auth/seams.py
@@ -1,23 +1,30 @@
 from __future__ import annotations
 
+from agendable.services.google_calendar_client import GoogleCalendarHttpClient
+from agendable.settings import get_settings
 from agendable.sso.oidc.client import OidcClient
 from agendable.web.routes import auth as auth_routes
 
 
 def oidc_enabled() -> bool:
-    enabled_fn = getattr(auth_routes, "_oidc_enabled", auth_routes.oidc_enabled)
-    return enabled_fn()
+    return auth_routes.oidc_enabled()
 
 
 def oidc_oauth_client() -> OidcClient:
-    client_fn = getattr(auth_routes, "_oidc_oauth_client", auth_routes.oidc_oauth_client)
-    return client_fn()
+    return auth_routes.oidc_oauth_client()
+
+
+def keycloak_oidc_enabled() -> bool:
+    return auth_routes.keycloak_oidc_enabled()
 
 
 def keycloak_oidc_oauth_client() -> OidcClient:
-    client_fn = getattr(
-        auth_routes,
-        "_keycloak_oidc_oauth_client",
-        auth_routes.keycloak_oidc_oauth_client,
+    return auth_routes.keycloak_oidc_oauth_client()
+
+
+def build_google_calendar_client() -> GoogleCalendarHttpClient:
+    settings = get_settings()
+    return GoogleCalendarHttpClient(
+        api_base_url=settings.google_calendar_api_base_url,
+        initial_sync_days_back=settings.google_calendar_initial_sync_days_back,
     )
-    return client_fn()
